requirements.txt
pandas
tqdm
argparse
openai
google-generativeai
pyyaml
python-dotenv